backend/requirements.txt
fastapi
uvicorn
supabase
imap-tools
python-dotenv